switches to dune and menhir
Since we have a circular dependency between lexer and parser the only
solution was to switch to menhir as it allows to split the grammar
into several modules and, in particular, to factor out tokens.

The switch required us to change the grammar definition as menhir is
treating the end of stream conflicts a little bit differently. The
introduced change might have unexpected consequences, though so we
need to add tests to make sure that everything still works more or
less as expected (we actually don't even know what to expect from this
parser).

Also, the indentation used in this code was pretty arbitrary so that
we had to rectify it a little bit to make code comprehensible. This
resulted in a huge diff.