Prepare 4.0.0 release (#16)
* repacks frontc into three independent packages

The printc command is no longer installed. The `ctoxml` and `calipso`
packages are distributed in separate packages.

* adds tests and more docs

* adds missing files

* adds github actions test

* runs the test only on pull requests

* copies the test artifacts to the test.t folder

to make windows happy, apparently links do not work on windows :)

* adds the missing menhir dependency

* removes deprecated names as they do not woron on macOS and Windows

The difference is in the register of letters, e.g., `Frontc` vs
`FrontC`, but it doesn't really affect the module names so there was
no need to introduce them at all.