adds a test to PR #19 (#20)
* adds a test to PR #19

I forgot to add so here it is. Most likely it will fail, not because
of the test but because dune will run the test in a wrong package,
where ctoxml is not yet available. But we will see.

* switch to dune 2.8 package stanza in cram

* adds applies_to to cram stanzas, removes build predicate for dune

* build and test directly instead of using opam

it looks like the cram packages do not work right with opam --with-tests