relaxes the bitfield declaration (#52)
Modern compilers accept bitfields with typedefed types despite that the
standard is not quite clear if this acceptable. FrontC strictly
requires a builtin integer type, e.g., `int`, `unsigned`, or `signed`
as the type of bitfiled. The reason for this is that we need to lookup
for the signedness of the field to construct a proper bitfield and
since we're not storing the original type of a typedefed type anywhere
we can't get this information for a typedefed type. The quick fix is
to keep signedness unspecified for the typedefed types.

Fixes #51